requirements.txt
torch>=1.13.0
transformers>=4.30.0
datasets>=2.14.0
rouge-score>=0.1.2
numpy>=1.21.0
pyyaml>=6.0